% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tokens.R
\name{count_tokens}
\alias{count_tokens}
\title{Count Tokens}
\usage{
count_tokens(data, group = NA, text = "text")
}
\arguments{
\item{data}{Data frame containing the raw data}

\item{group}{The name of the column(s) containing the grouping variable. If not defined, the text will not be grouped. Can be given as either a string or a vector of strings.}

\item{text}{The name of the column containing the text that needs to be tokenized}
}
\value{
Data frame containing columns for the word, the group(s) and the count labeled as 'word', the group name, and 'n'
}
\description{
Converts a data frame with columns for text and grouping variables into a data frame with each word and the count of each word in each group.
}
\examples{
test = data.frame (
  myText = c(
    "Hello! This is the first sentence I am using to test this function!",
    "This is the second sentence!"
  ),
  myGroup = c(
    "group1",
    "group2"
  )
)

count_tokens(test, text = "myText", group = "myGroup")
}
