\name{dhmeasures}
\alias{dhmeasures}
\alias{dhmeasures}
\docType{package}
\title{
  Digital History Measures
}
\description{
  Provides statistical functions to aid in the analysis of historical corpuses. The current functions include log likelihood (log_likelihood) and JSD (jsd).
}
\details{
  
}
\author{
  Ryan Schaefer
}
\keyword{ package }