% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_1820_example.R
\docType{data}
\name{hansard_1820_example}
\alias{hansard_1820_example}
\title{The 19th-century British Parliamentary debates for the decade 1820}
\format{
A data frame with 1082539 rows and 3 variables:

speaker
The name of the speaker originally recorded in the transciptions of the debates.

word
A word spoken by the given speaker.

n
The number of times the given word was spoken by the given speaker
}
\source{
{../data/hansard_1820_example.RData}
}
\usage{
hansard_1820_example
}
\description{
Hansard corpus data for the decade 1820. This data has been pre-formatted to contain words counts by speaker. Stopwords have been removed from the data. To access the raw Hansard data, install the package hansardr.
The variables are as follows:
}
\examples{
data(hansard_1820_example)

}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
}
\keyword{datasets}
