% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log Likelihood}
\usage{
log_likelihood(
  text,
  group_list = as.character(c()),
  word_list = as.character(c()),
  group = "group",
  word = "word",
  n = "n"
)
}
\arguments{
\item{text}{Data frame containing data}

\item{group_list}{Vector containing all groups to find log likelihood scores for}

\item{word_list}{Vector containing all words to find log likelihood scores for}

\item{group}{Name of data frame column containing text group}

\item{word}{Name of data frame column containing words}

\item{n}{Name of data frame column containing word count in text group}
}
\value{
Data frame containing a column containing unique words and columns for log likelihood scores for each group
}
\description{
Calculates word distinctiveness using the log likelihood algorithm. You input a data frame with columns for the word, the text group, and the number of times that word appears in that group. The column names are set to "word", "group", and "n" by default but they can be changed using the parameters word, group, and n. If any of these columns are not found, the function will not work. The output will be a new data frame with a column called "word" containing all unique words and subsequent columns for all unique groups with the name of that group. The data frame will contain the log likelihood scores for each word in each group. The larger a log likelihood score is, the more distinctive that word is to that group.
}
\examples{
\dontrun{
# Load example Hansard 1820 dataset
data(hansard_1820_example)
head(hansard_1820_example)

# Compute log likelihood
output = log_likelihood(
  hansard_1820_example, 
  group = "speaker", 
  group_list = c("Mr. Hume", "Mr. Brougham"), 
  word_list = c("house", "person")
)
head(output)
}
}
