% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stop_words.R
\name{remove_stop_words}
\alias{remove_stop_words}
\title{Remove Stop Words}
\usage{
remove_stop_words(data, words = "word", stop_words = dhmeasures::stop_word)
}
\arguments{
\item{data}{Data frame containing your data}

\item{words}{The name of the column where stop words should be searched for}

\item{stop_words}{Vector of stop words. Uses dhmeasures::stop_word as the default}
}
\value{
Data frame with all prior data, but without rows containing stop words
}
\description{
Remove stop words and fully numeric words from a column in a dataframe that contains words. It is assumed that the text has been tokenized (see dhmeasures::tokenize_counts) prior to using this function.
}
\examples{
# Load example Hansard 1820 dataset
data(hansard_1820_example)
head(hansard_1820_example)

head(remove_stop_words(hansard_1820_example))
}
