% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_word.R
\docType{data}
\name{stop_word}
\alias{stop_word}
\title{Stop Words List}
\format{
A character vector with 478 values
}
\source{
{../data/stop_word.RData}
}
\usage{
stop_word
}
\description{
A list of common words (known as stop words) that can be used to remove insignificant words from your data.
}
\examples{
data(stop_word)

}
\references{
Buongiorno, Steph (2021)
}
\keyword{datasets}
