\name{dhmeasures}
\alias{dhmeasures}
\alias{dhmeasures}
\docType{package}
\title{
  Digital History Measures
}
\description{
  Provides statistical functions to aid in the analysis of historical corpuses.
}
\details{
  The current functions include Log Likelihood (log_likelihood), JSD (jsd) and Partial JSD (partial_jsd).
}
\author{
  Steph Buongiorno and Ryan Schaefer
}
\keyword{ package }