\name{dhmeasures}
\alias{dhmeasures}
\alias{dhmeasures}
\docType{package}
\title{
  Digital History Measures
}
\description{
  Provides statistical and other helper functions to aid in the analysis of historical corpuses.
}
\details{
  The current statistical functions include Log Likelihood (log_likelihood), JSD (jsd) and Partial JSD (partial_jsd). The current helper functions include tokenize_counts.
}
\author{
  Steph Buongiorno and Ryan Schaefer
}
\keyword{ package }